feat(financialaid): classify how far each SSBTEK answer can be trusted, and stop truncating days
Two defects, both in the SO (Sveriges A-kassors Samorganisation) path.

1. StatusSvarandeOrganisation was read by nobody. The extractor walked
   Arbetsloshetsersattning(*) -> Utbetalningar(*) and never looked at the status, so an
   a-kassa answering "I cannot answer" with an empty payment list was indistinguishable
   from one answering "this person has nothing with us". Both produced silence, the
   normberakning showed no a-kassa income, and nothing flagged it.

   extractAnswers() now reads each responding organisation's status plus the structural
   signals the contract does define (AnsoktOmErsattning is 0..1, Utbetalningar is 0..n).
   Decision_ssbtekSvarKvalitet classifies it as SVARAT / INGET_HOS_OSS / EJ_KONTROLLERBAR,
   and an unverifiable answer surfaces to the handlaggare through incomeUnhandled as
   "A-kassa <org>: kunde inte kontrolleras" - a statement about our check, never about
   the sokande.

   The codes are not interpreted in Java on purpose: the contract declares the field
   xs:string with no enumeration and no annotation, so its codes are undefined by the
   contract itself, and we have no contact to ask. Putting the reading in a runtime DMN
   means the day someone establishes the codes it is one row and a publish, no deploy.
   Default is EJ_KONTROLLERBAR, including when the table is missing or the engine refuses
   it - an unavailable regelverk is not a clean bill of health.

2. Fractional days were truncated. Ersattningsdagar is xs:decimal in the contract and FK
   sends partial parental-benefit days, but both went through BigDecimal.intValue(), so
   4.5 days became 4 and 0.5 became none - before any rule got to decide what half a day
   means. SsbtekIncome.days is now BigDecimal.

Reasoning and what changes when the codes are established:
vof-ekonomiskt-bistand/architecture/ssbtek-svarskvalitet.md

+++ operaton-workers-financialaid/src/main/java/se/sundsvall/operaton/workers/financialaid/rules/SsbtekIncome.java
@@ -8,6 +8,10 @@ import java.time.LocalDate;
  * One normalised SSBTEK income, parsed from the api-service-financial-aid basis. The rule DMNs key off benefit,
  * sub-benefit, and amount type; {@code netAmount} and {@code period} drive period selection and change detection. The
  * serialized JSON keys stay aligned with the downstream Swedish contract.
+ * <p>
+ * {@code days} is a decimal, not a whole number: the SO contract declares {@code Ersattningsdagar} as
+ * {@code xs:decimal} and FK sends partial parental-benefit days, so quarter and half days are native to both formats.
+ * Truncating them to an int silently turned half a day into none, before any rule got to decide what half a day means.
  */
 public record SsbtekIncome(
 	@JsonProperty("forman") String benefit,
@@ -17,7 +21,7 @@ public record SsbtekIncome(
 	LocalDate period,
 	@JsonProperty("periodFran") LocalDate periodFrom,
 	@JsonProperty("periodTill") LocalDate periodTo,
-	@JsonProperty("dagar") Integer days,
+	@JsonProperty("dagar") BigDecimal days,
 	ApplicantRole role) {
 
 	/**
